Rework egl image api
