+++ src/wayland/mod.rs
@@ -19,7 +19,6 @@
 //!   quickly encounter a panic.
 
 pub mod compositor;
-//pub mod drm;
 pub mod output;
 pub mod seat;
 pub mod shm;
